PopulationGraphFrameコンポーネント内の関数をカスタムフックで管理
