+++ src/components/organisms/PopulationGraphFrame/PopulationGraphFrame.tsx
@@ -1,38 +1,16 @@
-import { useState, useEffect } from 'react';
+import { useState } from 'react';
 import Typography from '@/components/atoms/Typography/Typography';
 import PopulationGraph from '@/components/molecules/PopulationGraph/PopulationGraph';
 import PopulationTabGroup from '@/components/molecules/PopulationTabGroup/PopulationTabGroup';
 import { populationTabDefs } from '@/consts/PopulationTabDefs';
-import { PopulationDataWithPrefecture, PopulationTabId } from '@/types/resas';
+import usePopulationData from '@/hooks/usePopulationData';
+import { PopulationTabId } from '@/types/resas';
 import styles from './PopulationGraphFrame.module.scss';
 import { PrefectureGraphProps } from './PopulationGraphFrame.types';
 
 const PopulationGraphFrame: React.FC<PrefectureGraphProps> = ({ prefectures }) => {
-  const [populationData, setPopulationData] = useState<PopulationDataWithPrefecture[]>([]);
+  const populationData = usePopulationData(prefectures);
   const [selectedTab, setSelectedTab] = useState<PopulationTabId>(PopulationTabId.TotalPopulation);
-
-  useEffect(() => {
-    const getData = async () => {
-      try {
-        const response = await fetch('/api/populationData', {
-          method: 'POST',
-          headers: { 'Content-Type': 'application/json' },
-          body: JSON.stringify({ prefectures }),
-        });
-        if (!response.ok) {
-          throw new Error('Could not get data');
-        }
-        const data: PopulationDataWithPrefecture[] = await response.json();
-        setPopulationData(data);
-      } catch (error: unknown) {
-        setPopulationData([]);
-      }
-    };
-
-    if (prefectures.length > 0) {
-      getData();
-    }
-  }, [prefectures]);
 
   return (
     <div className={styles.container}>
